# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AuroraShield Security API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AnalyzePromptBody {
  /** The text to analyze for prompt injection */
  text: string;
  /** The site context (e.g. chatgpt, gmail, outlook) */
  context?: string;
}

export type PromptAnalysisResultRiskLevel =
  (typeof PromptAnalysisResultRiskLevel)[keyof typeof PromptAnalysisResultRiskLevel];

export const PromptAnalysisResultRiskLevel = {
  safe: "safe",
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface PromptAnalysisResult {
  /** Risk score from 0.0 (safe) to 1.0 (critical) */
  riskScore: number;
  riskLevel: PromptAnalysisResultRiskLevel;
  /** Detected injection patterns */
  patterns: string[];
  recommendation: string;
  cached: boolean;
}

export interface AnalyzePhishingBody {
  /** Page or email content to scan */
  content: string;
  /** The page URL for context */
  url?: string;
}

export type PhishingAnalysisResultRiskLevel =
  (typeof PhishingAnalysisResultRiskLevel)[keyof typeof PhishingAnalysisResultRiskLevel];

export const PhishingAnalysisResultRiskLevel = {
  safe: "safe",
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface PhishingAnalysisResult {
  riskScore: number;
  riskLevel: PhishingAnalysisResultRiskLevel;
  /** Suspicious phrases found in the content */
  suspiciousPhrases: string[];
  /** Phishing indicator types detected */
  indicators: string[];
  cached: boolean;
}

export interface AnalyzeUrlBody {
  /** List of URLs to analyze */
  urls: string[];
  /** The current page URL */
  pageUrl?: string;
}

export type UrlAnalysisResultRiskLevel =
  (typeof UrlAnalysisResultRiskLevel)[keyof typeof UrlAnalysisResultRiskLevel];

export const UrlAnalysisResultRiskLevel = {
  safe: "safe",
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface SingleUrlResult {
  url: string;
  riskScore: number;
  flags: string[];
}

export interface UrlAnalysisResult {
  overallRiskScore: number;
  riskLevel: UrlAnalysisResultRiskLevel;
  urlResults: SingleUrlResult[];
  cached: boolean;
}

export type AnalysisStatsThreatsByLevel = {
  safe: number;
  low: number;
  medium: number;
  high: number;
  critical: number;
};

export interface AnalysisStats {
  totalScans: number;
  promptInjections: number;
  phishingDetected: number;
  suspiciousUrls: number;
  threatsByLevel: AnalysisStatsThreatsByLevel;
}

export type ThreatEventType =
  (typeof ThreatEventType)[keyof typeof ThreatEventType];

export const ThreatEventType = {
  prompt_injection: "prompt_injection",
  phishing: "phishing",
  suspicious_url: "suspicious_url",
} as const;

export type ThreatEventRiskLevel =
  (typeof ThreatEventRiskLevel)[keyof typeof ThreatEventRiskLevel];

export const ThreatEventRiskLevel = {
  safe: "safe",
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface ThreatEvent {
  id: number;
  type: ThreatEventType;
  riskLevel: ThreatEventRiskLevel;
  riskScore: number;
  summary: string;
  context: string;
  detectedAt: string;
}