# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AuroraShield Security API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Checks user input for prompt injection patterns and LLM attacks
 * @summary Analyze text for prompt injection
 */
export const AnalyzePromptBody = zod.object({
  text: zod.string().describe("The text to analyze for prompt injection"),
  context: zod
    .string()
    .optional()
    .describe("The site context (e.g. chatgpt, gmail, outlook)"),
});

export const AnalyzePromptResponse = zod.object({
  riskScore: zod
    .number()
    .describe("Risk score from 0.0 (safe) to 1.0 (critical)"),
  riskLevel: zod.enum(["safe", "low", "medium", "high", "critical"]),
  patterns: zod.array(zod.string()).describe("Detected injection patterns"),
  recommendation: zod.string(),
  cached: zod.boolean(),
});

/**
 * Scans page content or email text for phishing indicators
 * @summary Analyze content for phishing
 */
export const AnalyzePhishingBody = zod.object({
  content: zod.string().describe("Page or email content to scan"),
  url: zod.string().optional().describe("The page URL for context"),
});

export const AnalyzePhishingResponse = zod.object({
  riskScore: zod.number(),
  riskLevel: zod.enum(["safe", "low", "medium", "high", "critical"]),
  suspiciousPhrases: zod
    .array(zod.string())
    .describe("Suspicious phrases found in the content"),
  indicators: zod
    .array(zod.string())
    .describe("Phishing indicator types detected"),
  cached: zod.boolean(),
});

/**
 * Checks URLs for domain length, suspicious patterns, and known threats
 * @summary Analyze URLs for suspicious patterns
 */
export const AnalyzeUrlBody = zod.object({
  urls: zod.array(zod.string()).describe("List of URLs to analyze"),
  pageUrl: zod.string().optional().describe("The current page URL"),
});

export const AnalyzeUrlResponse = zod.object({
  overallRiskScore: zod.number(),
  riskLevel: zod.enum(["safe", "low", "medium", "high", "critical"]),
  urlResults: zod.array(
    zod.object({
      url: zod.string(),
      riskScore: zod.number(),
      flags: zod.array(zod.string()),
    }),
  ),
  cached: zod.boolean(),
});

/**
 * Returns aggregate stats on threats detected across all sessions
 * @summary Get detection statistics
 */
export const GetAnalysisStatsResponse = zod.object({
  totalScans: zod.number(),
  promptInjections: zod.number(),
  phishingDetected: zod.number(),
  suspiciousUrls: zod.number(),
  threatsByLevel: zod.object({
    safe: zod.number(),
    low: zod.number(),
    medium: zod.number(),
    high: zod.number(),
    critical: zod.number(),
  }),
});

/**
 * Returns the most recently detected threats for display in the dashboard
 * @summary Get recent detected threats
 */
export const GetRecentThreatsResponseItem = zod.object({
  id: zod.number(),
  type: zod.enum(["prompt_injection", "phishing", "suspicious_url"]),
  riskLevel: zod.enum(["safe", "low", "medium", "high", "critical"]),
  riskScore: zod.number(),
  summary: zod.string(),
  context: zod.string(),
  detectedAt: zod.coerce.date(),
});
export const GetRecentThreatsResponse = zod.array(GetRecentThreatsResponseItem);